+++ libs/designer-ui/src/lib/settings/settingsection/settingTokenField.tsx
@@ -69,7 +69,9 @@ export const SettingTokenField = ({ ...props }: SettingTokenFieldProps) => {
           </Label>
         </div>
       )}
-      {isCustomEditor(props) ? <CustomTokenField {...props} labelId={labelId} /> : <TokenField {...props} labelId={labelId} />}
+      <div key={props.id}>
+        {isCustomEditor(props) ? <CustomTokenField {...props} labelId={labelId} /> : <TokenField {...props} labelId={labelId} />}
+      </div>
     </>
   );
 };
@@ -298,6 +300,7 @@ export const TokenField = ({
           supportedTypes={editorOptions?.supportedTypes}
           useStaticInputs={editorOptions?.useStaticInputs}
           initialValue={value}
+          isRequestApiConnectionTrigger={editorOptions?.isRequestApiConnectionTrigger}
           onChange={onValueChange}
         />
       );