+++ libs/services/designer-client-services/src/lib/__test__/connectionParameterEditor.spec.ts
@@ -1,0 +1,25 @@
+import {
+  ConnectionParameterEditorService,
+  InitConnectionParameterEditorService,
+  type IConnectionParameterInfo,
+} from '../connectionParameterEditor';
+
+describe('lib/designer-client-services/connectionParameterEditor', () => {
+  it('should call "getConnectionParameterEditor" with initialized service instance', () => {
+    const service = {
+      getConnectionParameterEditor: jest.fn(),
+    };
+
+    expect(ConnectionParameterEditorService()).toBeUndefined();
+
+    InitConnectionParameterEditorService(service);
+
+    const parameter: IConnectionParameterInfo = {
+      connectorId: 'connectorId',
+      parameterKey: 'parameterKey',
+    };
+    ConnectionParameterEditorService()?.getConnectionParameterEditor(parameter);
+
+    expect(service.getConnectionParameterEditor).toHaveBeenCalledWith(parameter);
+  });
+});